server/requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.14
django-cors-headers==3.10.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
polyline==1.4.0
psycopg2-binary==2.7.7
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.10
gunicorn==20.1.0